requirements.txt
# requirements.txt - Off-Road Segmentation Project
# GPU: NVIDIA RTX PRO 4000 Blackwell (sm_120)
# PyTorch는 setup.sh에서 cu128로 별도 설치

# === EfficientViT 핵심 의존성 ===
torchmetrics
numpy
timm
einops
opencv-python
tqdm
torchprofile
scipy
matplotlib

# === ONNX (Qualcomm IQ-9075 export용) ===
onnx
onnxsim
onnxruntime

# === 로깅 & 시각화 ===
wandb
tensorboard

# === 유틸리티 ===
pyyaml
Pillow
scikit-image
